@@ -1,11 +1,12 @@
 import os
 import tempfile
 from pathlib import Path
-from typing import Any, Union
+from typing import Any
 from unittest import mock
 
 import dask.array as da
 import numpy as np
+import pandas as pd
 import pytest
 import soundfile as sf
 from dask.array.core import Array as DaArray
@@ -311,7 +312,7 @@ class TestChannelFrame:
                 self.channel_frame, "compute", return_value=self.data
             ) as mock_compute:
                 mock_ax: mock.MagicMock = mock.MagicMock()
-                _: Union[Axes, Any] = self.channel_frame.plot(
+                _: Axes | Any = self.channel_frame.plot(
                     plot_type="waveform", ax=mock_ax
                 )
 
@@ -331,16 +332,6 @@ class TestChannelFrame:
         # Check that the data was reshaped
         assert cf.shape == (16000,)
         assert cf.n_channels == 1
-
-    def test_initialization_error_high_dim(self) -> None:
-        """Test initialization with data that has too many dimensions."""
-        data_3d = np.random.random((2, 16000, 3))
-        dask_data_3d = _da_from_array(data_3d, chunks=(1, 4000, 3))
-
-        with pytest.raises(
-            ValueError, match="Data must be 1-dimensional or 2-dimensional."
-        ):
-            ChannelFrame(dask_data_3d, self.sample_rate)
 
     def test_save_method(self) -> None:
         """Test saving audio to file."""
@@ -1520,25 +1511,31 @@ class TestDescribeIntegration:
 
     def test_visualize_graph(self) -> None:
         """Test visualize_graph method."""
-        # Test successful visualization
+        # Test successful visualization - returns mock object from visualize()
         with mock.patch.object(DaArray, "visualize") as mock_visualize:
-            filename = self.channel_frame.visualize_graph()
+            mock_return_value = mock.MagicMock()
+            mock_visualize.return_value = mock_return_value
+            result = self.channel_frame.visualize_graph()
             mock_visualize.assert_called_once()
-            assert filename is not None
+            # visualize_graph returns the result from _data.visualize()
+            assert result is mock_return_value
 
         # Test with provided filename
         with mock.patch.object(DaArray, "visualize") as mock_visualize:
+            mock_return_value = mock.MagicMock()
+            mock_visualize.return_value = mock_return_value
             custom_filename = "test_graph.png"
-            filename = self.channel_frame.visualize_graph(filename=custom_filename)
+            result = self.channel_frame.visualize_graph(filename=custom_filename)
             mock_visualize.assert_called_with(filename=custom_filename)
-            assert filename == custom_filename
+            # Returns the mock object from visualize()
+            assert result is mock_return_value
 
         # Test handling of visualization error
         with mock.patch.object(
             DaArray, "visualize", side_effect=Exception("Test error")
         ):
-            filename = self.channel_frame.visualize_graph()
-            assert filename is None
+            result = self.channel_frame.visualize_graph()
+            assert result is None
 
     def test_rms_property(self) -> None:
         """Test RMS property calculation."""
@@ -1728,3 +1725,541 @@ class TestBaseFrameExceptionHandling:
         with mock.patch.object(DaArray, "compute", return_value="not_an_array"):
             with pytest.raises(ValueError, match="Computed result is not a np.ndarray"):
                 _ = self.channel_frame.compute()
+
+    # Error Message Tests (Phase 1 Improvements)
+    def test_invalid_data_shape_error_message(self) -> None:
+        """Test that invalid data shape provides helpful error message."""
+        data_3d = np.random.random((2, 3, 4))  # 3D array
+        dask_data_3d = _da_from_array(data_3d)
+
+        with pytest.raises(ValueError) as exc_info:
+            ChannelFrame(data=dask_data_3d, sampling_rate=16000)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Invalid data shape" in error_msg
+        assert "(2, 3, 4)" in error_msg
+        assert "3D" in error_msg
+        # Check WHY
+        assert "Expected: 1D" in error_msg
+        assert "2D" in error_msg
+        # Check HOW
+        assert "reshape" in error_msg.lower()
+        assert "Example:" in error_msg
+
+    def test_negative_sampling_rate_error_message(self) -> None:
+        """Test that negative sampling rate provides helpful error message."""
+        with pytest.raises(ValueError) as exc_info:
+            ChannelFrame(data=self.dask_data, sampling_rate=-44100)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Invalid sampling_rate" in error_msg
+        assert "-44100" in error_msg
+        # Check WHY
+        assert "Positive value" in error_msg
+        # Check HOW
+        assert "Common values:" in error_msg
+        assert "44100" in error_msg
+
+    def test_zero_sampling_rate_error_message(self) -> None:
+        """Test that zero sampling rate provides helpful error message."""
+        with pytest.raises(ValueError) as exc_info:
+            ChannelFrame(data=self.dask_data, sampling_rate=0)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Invalid sampling_rate" in error_msg
+        assert "Got: 0 Hz" in error_msg
+        # Check WHY
+        assert "Expected: Positive value > 0" in error_msg
+        # Check HOW
+        assert (
+            "Sampling rate represents samples per second and must be positive."
+            in error_msg
+        )
+        assert "Common values: 8000, 16000, 22050, 44100, 48000 Hz" in error_msg
+
+    def test_file_not_found_error_message(self) -> None:
+        """Test that missing file provides helpful error message."""
+        fake_path = "/nonexistent/path/to/audio.wav"
+
+        with pytest.raises(FileNotFoundError) as exc_info:
+            ChannelFrame.from_file(fake_path)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Audio file not found" in error_msg
+        # Cross-platform: check for key path components instead of exact path
+        assert "nonexistent" in error_msg and "audio.wav" in error_msg
+        # Check WHY (context)
+        assert "Current directory:" in error_msg
+        # Check HOW
+        assert "check" in error_msg.lower()
+        assert "File path is correct" in error_msg
+        assert "File exists" in error_msg
+
+
+class TestFadeIntegration:
+    """Integration tests for fade functionality with other operations."""
+
+    def setup_method(self) -> None:
+        """Set up test fixtures for fade integration tests."""
+        # Create a test signal with known properties
+        self.sample_rate = 16000
+        duration = 1.0  # 1 second
+        n_samples = int(self.sample_rate * duration)
+        t = np.linspace(0, duration, n_samples, endpoint=False)
+
+        # Create a sine wave with amplitude 1.0
+        freq = 440  # Hz
+        signal = np.sin(2 * np.pi * freq * t)
+        self.data = signal.reshape(1, -1)  # Single channel
+        self.dask_data = _da_from_array(self.data, chunks=(1, 4000))
+        self.channel_frame = ChannelFrame(
+            data=self.dask_data, sampling_rate=self.sample_rate, label="test_sine"
+        )
+
+    def test_fade_preserves_rms_calculation(self) -> None:
+        """Test that fade operation allows RMS calculation to work correctly."""
+        # Apply fade
+        faded = self.channel_frame.fade(fade_ms=100.0)
+
+        # RMS should be calculable without errors
+        rms_values = faded.rms
+        assert len(rms_values) == 1
+        assert rms_values[0] > 0  # RMS should be positive
+
+        # Faded signal should have lower RMS than original (due to fade-in/out)
+        original_rms = self.channel_frame.rms
+        assert rms_values[0] < original_rms[0]
+
+    def test_fade_with_normalize_chain(self) -> None:
+        """Test fade operation chained with normalize."""
+        # Chain fade and normalize operations
+        processed = self.channel_frame.fade(fade_ms=50.0).normalize()
+
+        # Should complete without errors
+        assert isinstance(processed, ChannelFrame)
+        assert processed.sampling_rate == self.sample_rate
+        assert processed.n_channels == 1
+        assert processed.n_samples == self.channel_frame.n_samples
+
+        # Check that operation history is recorded
+        assert len(processed.operation_history) == 2
+        assert processed.operation_history[0]["operation"] == "fade"
+        assert processed.operation_history[1]["operation"] == "normalize"
+
+        # Normalized signal should have max amplitude of 1.0
+        max_amplitude = np.max(np.abs(processed.data))
+        np.testing.assert_almost_equal(max_amplitude, 1.0, decimal=6)
+
+    def test_fade_with_filter_chain(self) -> None:
+        """Test fade operation chained with filtering."""
+        # Chain fade and low-pass filter
+        processed = self.channel_frame.fade(fade_ms=50.0).low_pass_filter(cutoff=1000)
+
+        # Should complete without errors
+        assert isinstance(processed, ChannelFrame)
+        assert processed.sampling_rate == self.sample_rate
+
+        # Check operation history
+        assert len(processed.operation_history) == 2
+        assert processed.operation_history[0]["operation"] == "fade"
+        assert processed.operation_history[1]["operation"] == "lowpass_filter"
+
+    def test_fade_with_multiple_operations_chain(self) -> None:
+        """Test fade in a complex operation chain."""
+
+        # Create a more complex processing chain
+        processed = (
+            self.channel_frame.fade(fade_ms=25.0)
+            .normalize()
+            .low_pass_filter(cutoff=2000)
+            .high_pass_filter(cutoff=100)
+        )
+
+        # Should complete without errors
+        assert isinstance(processed, ChannelFrame)
+        assert processed.sampling_rate == self.sample_rate
+
+        # Check that all operations are recorded
+        assert len(processed.operation_history) == 4
+        operations = [op["operation"] for op in processed.operation_history]
+        assert operations == ["fade", "normalize", "lowpass_filter", "highpass_filter"]
+
+    def test_fade_with_channel_operations(self) -> None:
+        """Test fade with channel selection and operations."""
+        # Create multi-channel signal
+        multi_data = np.vstack([self.data[0], self.data[0] * 0.5])  # 2 channels
+        multi_dask = _da_from_array(multi_data, chunks=(1, 4000))
+        multi_frame = ChannelFrame(
+            data=multi_dask, sampling_rate=self.sample_rate, label="multi_test"
+        )
+
+        # Apply fade to all channels, then select one channel
+        processed = multi_frame.fade(fade_ms=50.0).get_channel(0)
+
+        # Should work correctly
+        assert isinstance(processed, ChannelFrame)
+        assert processed.n_channels == 1
+        assert processed.operation_history[-1]["operation"] == "fade"
+
+    def test_fade_with_arithmetic_operations(self) -> None:
+        """Test fade with arithmetic operations."""
+        # Apply fade, then add a constant
+        processed = self.channel_frame.fade(fade_ms=50.0) + 0.1
+
+        # Should complete without errors
+        assert isinstance(processed, ChannelFrame)
+        assert processed.operation_history[0]["operation"] == "fade"
+
+        # Check that arithmetic operation is recorded
+        assert "+" in str(processed.operation_history[1]["operation"])
+
+    def test_fade_preserves_metadata_and_labels(self) -> None:
+        """Test that fade preserves channel metadata and labels."""
+        # Set custom labels and metadata
+        self.channel_frame.channels[0].label = "test_channel"
+        self.channel_frame.channels[0]["gain"] = 0.8
+        self.channel_frame.metadata["test_key"] = "test_value"
+
+        # Apply fade
+        faded = self.channel_frame.fade(fade_ms=50.0)
+
+        # Check that label is updated to reflect the operation
+        assert faded.channels[0].label == "fade(test_channel)"
+        # Check that other metadata is preserved
+        assert faded.channels[0]["gain"] == 0.8
+        assert faded.metadata["test_key"] == "test_value"
+
+        # Check operation history
+        assert faded.operation_history[0]["operation"] == "fade"
+        assert faded.operation_history[0]["params"]["fade_ms"] == 50.0
+
+    def test_fade_with_file_io_roundtrip(self) -> None:
+        """Test fade operation with file save/load roundtrip."""
+        import os
+        import tempfile
+
+        with tempfile.NamedTemporaryFile(suffix=".wav", delete=False) as temp_file:
+            temp_filename = temp_file.name
+
+        try:
+            # Apply fade and save
+            faded = self.channel_frame.fade(fade_ms=50.0)
+            faded.to_wav(temp_filename)
+
+            # Load back and verify
+            loaded = ChannelFrame.from_file(temp_filename)
+
+            # Should be able to load and have same basic properties
+            assert loaded.sampling_rate == self.sample_rate
+            assert loaded.n_channels == 1
+            assert loaded.n_samples == self.channel_frame.n_samples
+
+            # Data should be different (faded) but same shape
+            assert loaded.data.shape == faded.data.shape
+
+        finally:
+            if os.path.exists(temp_filename):
+                os.remove(temp_filename)
+
+    def test_fade_with_different_fade_durations(self) -> None:
+        """Test fade with different fade durations and verify effects."""
+        # Test with very short fade
+        short_fade = self.channel_frame.fade(fade_ms=1.0)
+        short_rms = short_fade.rms[0]
+
+        # Test with longer fade
+        long_fade = self.channel_frame.fade(fade_ms=100.0)
+        long_rms = long_fade.rms[0]
+
+        # Longer fade should result in lower RMS (more signal attenuated)
+        assert long_rms < short_rms
+
+        # Both should be less than original
+        original_rms = self.channel_frame.rms[0]
+        assert short_rms < original_rms
+        assert long_rms < original_rms
+
+    def test_fade_with_multi_channel_signal(self) -> None:
+        """Test fade with multi-channel signal."""
+        # Create 3-channel signal
+        multi_data = np.vstack(
+            [
+                self.data[0],  # Original
+                self.data[0] * 0.7,  # Scaled down
+                self.data[0] * 1.3,  # Scaled up
+            ]
+        )
+        multi_dask = _da_from_array(multi_data, chunks=(1, 4000))
+        multi_frame = ChannelFrame(
+            data=multi_dask, sampling_rate=self.sample_rate, label="multi_test"
+        )
+
+        # Apply fade
+        faded = multi_frame.fade(fade_ms=50.0)
+
+        # Should work for all channels
+        assert faded.n_channels == 3
+        assert faded.operation_history[0]["operation"] == "fade"
+
+        # Each channel should have different RMS due to different amplitudes
+        rms_values = faded.rms
+        assert len(rms_values) == 3
+        assert rms_values[0] > rms_values[1]  # Original > scaled down
+        assert rms_values[2] > rms_values[0]  # Scaled up > original
+
+    def test_fade_lazy_evaluation_preserved(self) -> None:
+        """Test that fade preserves lazy evaluation."""
+        # Apply fade without computing
+        faded = self.channel_frame.fade(fade_ms=50.0)
+
+        # Should still be lazy (dask array)
+        assert isinstance(faded._data, DaArray)
+
+        # Operation history should be updated without computation
+        assert len(faded.operation_history) == 1
+        assert faded.operation_history[0]["operation"] == "fade"
+
+        # Only when we access .data should computation happen
+        with mock.patch.object(
+            DaArray, "compute", return_value=self.data
+        ) as mock_compute:
+            _ = faded.data
+            mock_compute.assert_called_once()
+
+    def test_fade_with_visualization(self) -> None:
+        """Test that faded signal can be visualized."""
+        # Apply fade
+        faded = self.channel_frame.fade(fade_ms=50.0)
+
+        # Should be able to create plots without errors
+        with (
+            mock.patch("matplotlib.pyplot.figure"),
+            mock.patch("matplotlib.pyplot.subplot"),
+            mock.patch("matplotlib.axes.Axes.plot"),
+            mock.patch("matplotlib.pyplot.tight_layout"),
+            mock.patch("matplotlib.pyplot.show"),
+        ):
+            # Basic plot should work
+            faded.plot()
+
+            # RMS plot should work
+            faded.rms_plot()
+
+    def test_fade_error_handling_integration(self) -> None:
+        """Test error handling in fade operation within processing chains."""
+        # Test with invalid fade_ms (too long)
+        # Create very short signal
+        short_data = self.data[:, :100]  # Only 100 samples
+        short_dask = _da_from_array(short_data, chunks=(1, 50))
+        short_frame = ChannelFrame(
+            data=short_dask, sampling_rate=self.sample_rate, label="short"
+        )
+
+        # Apply fade with duration longer than signal
+        # Should not fail immediately due to lazy eval
+        faded_short = short_frame.fade(fade_ms=10.0)
+
+        # Error should occur when we try to compute the result
+        with pytest.raises(ValueError, match="Fade length too long"):
+            _ = faded_short.data
+
+        # Test with negative fade_ms - fails during operation creation
+        with pytest.raises(ValueError, match="fade_ms must be non-negative"):
+            self.channel_frame.fade(fade_ms=-1.0)
+
+    def test_to_numpy(self) -> None:
+        """Test to_numpy method converts frame data to NumPy array."""
+        # For single channel, to_numpy should return 1D array (squeezed)
+        if self.data.shape[0] == 1:
+            expected_result = self.data.squeeze(axis=0)
+        else:
+            expected_result = self.data
+
+        # Test with mock to ensure compute is called
+        with mock.patch.object(
+            self.channel_frame, "compute", return_value=self.data
+        ) as mock_compute:
+            result = self.channel_frame.to_numpy()
+            mock_compute.assert_called_once()
+            np.testing.assert_array_equal(result, expected_result)
+            assert isinstance(result, np.ndarray)
+
+    def test_to_dataframe(self) -> None:
+        """Test to_dataframe method converts frame data to pandas DataFrame."""
+        # Test basic conversion
+        df = self.channel_frame.to_dataframe()
+
+        # Check DataFrame properties
+        assert isinstance(df, pd.DataFrame)
+        assert df.shape == (
+            self.data.shape[1],
+            self.data.shape[0],
+        )  # (n_samples, n_channels)
+        assert list(df.columns) == ["ch0"]  # Single channel
+        assert df.index.name == "time"
+
+        # Check data values (transposed)
+        np.testing.assert_array_equal(df.values.T, self.data)
+
+        # Check time index
+        expected_time = np.arange(self.data.shape[1]) / self.sample_rate
+        np.testing.assert_array_equal(df.index.values, expected_time)
+
+    def test_to_dataframe_with_custom_labels(self) -> None:
+        """Test to_dataframe with custom channel labels."""
+        # Set custom labels
+        self.channel_frame.channels[0].label = "left"
+
+        df = self.channel_frame.to_dataframe()
+
+        # Check column names
+        assert list(df.columns) == ["left"]
+
+    def test_to_dataframe_single_channel(self) -> None:
+        """Test to_dataframe with single channel."""
+        # Get single channel (already single channel)
+        single_channel = self.channel_frame.get_channel(0)
+
+        df = single_channel.to_dataframe()
+
+        # Check DataFrame properties
+        assert df.shape == (self.data.shape[1], 1)  # (n_samples, 1)
+        assert list(df.columns) == ["ch0"]
+        assert df.index.name == "time"
+
+        # Check data values
+        np.testing.assert_array_equal(df.values.flatten(), self.data[0])
+
+    def test_info_method_basic(self, capsys: Any) -> None:
+        """Test info() method displays correct information."""
+        self.channel_frame.info()
+
+        captured = capsys.readouterr()
+        output = captured.out
+
+        # Verify all expected information is present
+        assert "Channels: 1" in output
+        assert f"Sampling rate: {self.sample_rate} Hz" in output
+        assert "Duration: 1.0 s" in output
+        assert f"Samples: {self.channel_frame.n_samples}" in output
+        assert "Channel labels: ['ch0']" in output
+
+    def test_info_method_single_channel(self, capsys: Any) -> None:
+        """Test info() method with single channel."""
+        single_data = self.data[0:1]
+        single_frame = ChannelFrame.from_numpy(
+            single_data, sampling_rate=self.sample_rate, label="single"
+        )
+
+        single_frame.info()
+
+        captured = capsys.readouterr()
+        output = captured.out
+
+        assert "Channels: 1" in output
+        assert "Channel labels: ['ch0']" in output
+
+    def test_info_method_custom_labels(self, capsys: Any) -> None:
+        """Test info() method with custom channel labels."""
+        # Set custom labels
+        self.channel_frame.channels[0].label = "left"
+
+        self.channel_frame.info()
+
+        captured = capsys.readouterr()
+        output = captured.out
+
+        assert "Channel labels: ['left']" in output
+
+    def test_info_method_different_duration(self, capsys: Any) -> None:
+        """Test info() method with different durations."""
+        # Create a frame with 0.5 seconds of data
+        short_data = np.random.random((1, 8000))
+        short_frame = ChannelFrame.from_numpy(
+            short_data, sampling_rate=self.sample_rate, label="short"
+        )
+
+        short_frame.info()
+
+        captured = capsys.readouterr()
+        output = captured.out
+
+        assert "Duration: 0.5 s" in output
+        assert "Samples: 8000" in output
+
+    # Error Message Tests (Phase 1 Improvements)
+    def test_invalid_data_shape_error_message(self) -> None:
+        """Test that invalid data shape provides helpful error message."""
+        data_3d = np.random.random((2, 3, 4))  # 3D array
+        dask_data_3d = _da_from_array(data_3d)
+
+        with pytest.raises(ValueError) as exc_info:
+            ChannelFrame(data=dask_data_3d, sampling_rate=16000)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Invalid data shape" in error_msg
+        assert "(2, 3, 4)" in error_msg
+        assert "3D" in error_msg
+        # Check WHY
+        assert "Expected: 1D" in error_msg
+        assert "2D" in error_msg
+        # Check HOW
+        assert "reshape" in error_msg.lower()
+        assert "Example:" in error_msg
+
+    def test_negative_sampling_rate_error_message(self) -> None:
+        """Test that negative sampling rate provides helpful error message."""
+        with pytest.raises(ValueError) as exc_info:
+            ChannelFrame(data=self.dask_data, sampling_rate=-44100)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Invalid sampling_rate" in error_msg
+        assert "Got: -44100 Hz" in error_msg
+        # Check WHY
+        assert "Expected: Positive value > 0" in error_msg
+        # Check HOW
+        assert "Common values:" in error_msg
+        assert "44100" in error_msg
+
+    def test_zero_sampling_rate_error_message(self) -> None:
+        """Test that zero sampling rate provides helpful error message."""
+        with pytest.raises(ValueError) as exc_info:
+            ChannelFrame(data=self.dask_data, sampling_rate=0)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Invalid sampling_rate" in error_msg
+        assert "Got: 0 Hz" in error_msg
+        # Check WHY
+        assert "Expected: Positive value > 0" in error_msg
+        # Check HOW
+        assert (
+            "Sampling rate represents samples per second and must be positive."
+            in error_msg
+        )
+        assert "Common values: 8000, 16000, 22050, 44100, 48000 Hz" in error_msg
+
+    def test_file_not_found_error_message(self) -> None:
+        """Test that missing file provides helpful error message."""
+        fake_path = "/nonexistent/path/to/audio.wav"
+
+        with pytest.raises(FileNotFoundError) as exc_info:
+            ChannelFrame.from_file(fake_path)
+
+        error_msg = str(exc_info.value)
+        # Check WHAT
+        assert "Audio file not found" in error_msg
+        # Cross-platform: check for key path components instead of exact path
+        assert "nonexistent" in error_msg and "audio.wav" in error_msg
+        # Check WHY (context)
+        assert "Current directory:" in error_msg
+        # Check HOW
+        assert "check" in error_msg.lower()
+        assert "File path is correct" in error_msg
+        assert "File exists" in error_msg
