@@ -3,13 +3,16 @@ import logging
 import numbers
 import uuid
 from abc import ABC, abstractmethod
-from collections.abc import Iterator
-from typing import Any, Callable, Generic, Optional, TypeVar, Union, cast
+from collections.abc import Callable, Iterator
+from typing import Any, Generic, Optional, TypeVar, cast
 
 import numpy as np
 import numpy.typing as npt
+import pandas as pd
 from dask.array.core import Array as DaArray
+from IPython.display import Image as IPythonImage
 from matplotlib.axes import Axes
+from pydantic import ValidationError
 
 from wandas.utils.types import NDArrayComplex, NDArrayReal
 
@@ -41,8 +44,9 @@ class BaseFrame(ABC, Generic[T]):
         Additional metadata for the frame.
     operation_history : list[dict], optional
         History of operations performed on this frame.
-    channel_metadata : list[ChannelMetadata], optional
-        Metadata for each channel in the frame.
+    channel_metadata : list[ChannelMetadata | dict], optional
+        Metadata for each channel in the frame. Can be ChannelMetadata objects
+        or dicts that will be validated by Pydantic.
     previous : BaseFrame, optional
         The frame that this frame was derived from.
 
@@ -70,10 +74,10 @@ class BaseFrame(ABC, Generic[T]):
         self,
         data: DaArray,
         sampling_rate: float,
-        label: Optional[str] = None,
-        metadata: Optional[dict[str, Any]] = None,
-        operation_history: Optional[list[dict[str, Any]]] = None,
-        channel_metadata: Optional[list[ChannelMetadata]] = None,
+        label: str | None = None,
+        metadata: dict[str, Any] | None = None,
+        operation_history: list[dict[str, Any]] | None = None,
+        channel_metadata: list[ChannelMetadata] | list[dict[str, Any]] | None = None,
         previous: Optional["BaseFrame[Any]"] = None,
     ):
         self._data = data.rechunk(chunks=-1)  # type: ignore [unused-ignore]
@@ -86,7 +90,35 @@ class BaseFrame(ABC, Generic[T]):
         self._previous = previous
 
         if channel_metadata:
-            self._channel_metadata = copy.deepcopy(channel_metadata)
+            # Pydantic handles both ChannelMetadata objects and dicts
+            def _to_channel_metadata(
+                ch: ChannelMetadata | dict[str, Any], index: int
+            ) -> ChannelMetadata:
+                if isinstance(ch, ChannelMetadata):
+                    return copy.deepcopy(ch)
+                elif isinstance(ch, dict):
+                    try:
+                        return ChannelMetadata(**ch)
+                    except ValidationError as e:
+                        raise ValueError(
+                            f"Invalid channel_metadata at index {index}\n"
+                            f"  Got: {ch}\n"
+                            f"  Validation error: {e}\n"
+                            f"Ensure all dict keys match ChannelMetadata fields "
+                            f"(label, unit, ref, extra) and have correct types."
+                        ) from e
+                else:
+                    raise TypeError(
+                        f"Invalid type in channel_metadata at index {index}\n"
+                        f"  Got: {type(ch).__name__} ({ch!r})\n"
+                        f"  Expected: ChannelMetadata or dict\n"
+                        f"Use ChannelMetadata objects or dicts with valid fields."
+                    )
+
+            self._channel_metadata = [
+                _to_channel_metadata(cast(ChannelMetadata | dict[str, Any], ch), i)
+                for i, ch in enumerate(channel_metadata)
+            ]
         else:
             self._channel_metadata = [
                 ChannelMetadata(label=f"ch{i}", unit="", extra={})
@@ -126,13 +158,11 @@ class BaseFrame(ABC, Generic[T]):
 
     def get_channel(
         self: S,
-        channel_idx: Union[
-            int,
-            list[int],
-            tuple[int, ...],
-            npt.NDArray[np.int_],
-            npt.NDArray[np.bool_],
-        ],
+        channel_idx: int
+        | list[int]
+        | tuple[int, ...]
+        | npt.NDArray[np.int_]
+        | npt.NDArray[np.bool_],
     ) -> S:
         """
         Get channel(s) by index.
@@ -181,27 +211,23 @@ class BaseFrame(ABC, Generic[T]):
 
     def __getitem__(
         self: S,
-        key: Union[
-            int,
-            str,
-            slice,
-            list[int],
-            list[str],
-            tuple[
-                Union[
-                    int,
-                    str,
-                    slice,
-                    list[int],
-                    list[str],
-                    npt.NDArray[np.int_],
-                    npt.NDArray[np.bool_],
-                ],
-                ...,
-            ],
-            npt.NDArray[np.int_],
-            npt.NDArray[np.bool_],
-        ],
+        key: int
+        | str
+        | slice
+        | list[int]
+        | list[str]
+        | tuple[
+            int
+            | str
+            | slice
+            | list[int]
+            | list[str]
+            | npt.NDArray[np.int_]
+            | npt.NDArray[np.bool_],
+            ...,
+        ]
+        | npt.NDArray[np.int_]
+        | npt.NDArray[np.bool_],
     ) -> S:
         """
         Get channel(s) by index, label, or advanced indexing.
@@ -270,7 +296,8 @@ class BaseFrame(ABC, Generic[T]):
 
         # Single index (int)
         if isinstance(key, numbers.Integral):
-            return self.get_channel(key)
+            # Ensure we pass a plain Python int to satisfy the type checker
+            return self.get_channel(int(key))
 
         # Single label (str)
         if isinstance(key, str):
@@ -309,7 +336,7 @@ class BaseFrame(ABC, Generic[T]):
                 return self.get_channel(indices_from_labels)
 
             # Check if all elements are integers
-            elif all(isinstance(k, (int, np.integer)) for k in key):
+            elif all(isinstance(k, int | np.integer) for k in key):
                 # Multiple indices - convert to list[int] for type safety
                 int_list = [int(k) for k in key]
                 return self.get_channel(int_list)
@@ -344,15 +371,13 @@ class BaseFrame(ABC, Generic[T]):
     def _handle_multidim_indexing(
         self: S,
         key: tuple[
-            Union[
-                int,
-                str,
-                slice,
-                list[int],
-                list[str],
-                npt.NDArray[np.int_],
-                npt.NDArray[np.bool_],
-            ],
+            int
+            | str
+            | slice
+            | list[int]
+            | list[str]
+            | npt.NDArray[np.int_]
+            | npt.NDArray[np.bool_],
             ...,
         ],
     ) -> S:
@@ -383,9 +408,9 @@ class BaseFrame(ABC, Generic[T]):
         time_keys = key[1:] if len(key) > 1 else ()
 
         # Select channels first (recursively call __getitem__)
-        if isinstance(channel_key, (list, np.ndarray)):
+        if isinstance(channel_key, list | np.ndarray):
             selected = self[channel_key]
-        elif isinstance(channel_key, (int, str, slice)):
+        elif isinstance(channel_key, int | str | slice):
             selected = self[channel_key]
         else:
             raise TypeError(
@@ -478,8 +503,8 @@ class BaseFrame(ABC, Generic[T]):
 
     @abstractmethod
     def plot(
-        self, plot_type: str = "default", ax: Optional[Axes] = None, **kwargs: Any
-    ) -> Union[Axes, Iterator[Axes]]:
+        self, plot_type: str = "default", ax: Axes | None = None, **kwargs: Any
+    ) -> Axes | Iterator[Axes]:
         """Plot the data"""
         pass
 
@@ -541,12 +566,57 @@ class BaseFrame(ABC, Generic[T]):
             return result.astype(dtype)
         return result
 
-    def visualize_graph(self, filename: Optional[str] = None) -> Optional[str]:
-        """Visualize the computation graph and save it to a file"""
+    def visualize_graph(self, filename: str | None = None) -> IPythonImage | None:
+        """
+        Visualize the computation graph and save it to a file.
+
+        This method creates a visual representation of the Dask computation graph.
+        In Jupyter notebooks, it returns an IPython.display.Image object that
+        will be displayed inline. In other environments, it saves the graph to
+        a file and returns None.
+
+        Parameters
+        ----------
+        filename : str, optional
+            Output filename for the graph image. If None, a unique filename
+            is generated using UUID. The file is saved in the current working
+            directory.
+
+        Returns
+        -------
+        IPython.display.Image or None
+            In Jupyter environments: Returns an IPython.display.Image object
+            that can be displayed inline.
+            In other environments: Returns None after saving the graph to file.
+            Returns None if visualization fails.
+
+        Notes
+        -----
+        This method requires graphviz to be installed on your system:
+        - Ubuntu/Debian: `sudo apt-get install graphviz`
+        - macOS: `brew install graphviz`
+        - Windows: Download from https://graphviz.org/download/
+
+        The graph displays operation names (e.g., 'normalize', 'lowpass_filter')
+        making it easier to understand the processing pipeline.
+
+        Examples
+        --------
+        >>> import wandas as wd
+        >>> signal = wd.read_wav("audio.wav")
+        >>> processed = signal.normalize().low_pass_filter(cutoff=1000)
+        >>> # In Jupyter: displays graph inline
+        >>> processed.visualize_graph()
+        >>> # Save to specific file
+        >>> processed.visualize_graph("my_graph.png")
+
+        See Also
+        --------
+        debug_info : Print detailed debug information about the frame
+        """
         try:
             filename = filename or f"graph_{uuid.uuid4().hex[:8]}.png"
-            self._data.visualize(filename=filename)
-            return filename
+            return self._data.visualize(filename=filename)
         except Exception as e:
             logger.warning(f"Failed to visualize the graph: {e}")
             return None
@@ -554,7 +624,7 @@ class BaseFrame(ABC, Generic[T]):
     @abstractmethod
     def _binary_op(
         self: S,
-        other: Union[S, int, float, NDArrayReal, DaArray],
+        other: S | int | float | NDArrayReal | DaArray,
         op: Callable[[DaArray, Any], DaArray],
         symbol: str,
     ) -> S:
@@ -563,21 +633,25 @@ class BaseFrame(ABC, Generic[T]):
         # Actual implementation is left to derived classes
         pass
 
-    def __add__(self: S, other: Union[S, int, float, NDArrayReal]) -> S:
+    def __add__(self: S, other: S | int | float | NDArrayReal) -> S:
         """Addition operator"""
         return self._binary_op(other, lambda x, y: x + y, "+")
 
-    def __sub__(self: S, other: Union[S, int, float, NDArrayReal]) -> S:
+    def __sub__(self: S, other: S | int | float | NDArrayReal) -> S:
         """Subtraction operator"""
         return self._binary_op(other, lambda x, y: x - y, "-")
 
-    def __mul__(self: S, other: Union[S, int, float, NDArrayReal]) -> S:
+    def __mul__(self: S, other: S | int | float | NDArrayReal) -> S:
         """Multiplication operator"""
         return self._binary_op(other, lambda x, y: x * y, "*")
 
-    def __truediv__(self: S, other: Union[S, int, float, NDArrayReal]) -> S:
+    def __truediv__(self: S, other: S | int | float | NDArrayReal) -> S:
         """Division operator"""
         return self._binary_op(other, lambda x, y: x / y, "/")
+
+    def __pow__(self: S, other: S | int | float | NDArrayReal) -> S:
+        """Power operator"""
+        return self._binary_op(other, lambda x, y: x**y, "**")
 
     def apply_operation(self: S, operation_name: str, **params: Any) -> S:
         """
@@ -603,6 +677,52 @@ class BaseFrame(ABC, Generic[T]):
         """Implementation of operation application"""
         pass
 
+    def _relabel_channels(
+        self,
+        operation_name: str,
+        display_name: str | None = None,
+    ) -> list[ChannelMetadata]:
+        """
+        Update channel labels to reflect applied operation.
+
+        This method creates new channel metadata with labels that include
+        the operation name, making it easier to track processing history
+        and distinguish frames in plots.
+
+        Parameters
+        ----------
+        operation_name : str
+            Name of the operation (e.g., "normalize", "lowpass_filter")
+        display_name : str, optional
+            Display name for the operation. If None, uses operation_name.
+            This allows operations to provide custom, more readable labels.
+
+        Returns
+        -------
+        list[ChannelMetadata]
+            New channel metadata with updated labels.
+            Original metadata is deep-copied and only labels are modified.
+
+        Examples
+        --------
+        >>> # Original label: "ch0"
+        >>> # After normalize: "normalize(ch0)"
+        >>> # After chained ops: "lowpass_filter(normalize(ch0))"
+
+        Notes
+        -----
+        Labels are nested for chained operations, allowing full
+        traceability of the processing pipeline.
+        """
+        display = display_name or operation_name
+        new_metadata = []
+        for ch in self._channel_metadata:
+            # All channel metadata are ChannelMetadata objects at this point
+            new_ch = ch.model_copy(deep=True)
+            new_ch.label = f"{display}({ch.label})"
+            new_metadata.append(new_ch)
+        return new_metadata
+
     def debug_info(self) -> None:
         """Output detailed debug information"""
         logger.debug(f"=== {self.__class__.__name__} Debug Info ===")
@@ -613,6 +733,127 @@ class BaseFrame(ABC, Generic[T]):
         self._debug_info_impl()
         logger.debug("=== End Debug Info ===")
 
+    def print_operation_history(self) -> None:
+        """
+        Print the operation history to standard output in a readable format.
+
+        This method writes a human-friendly representation of the
+        `operation_history` list to stdout. Each operation is printed on its
+        own line with an index, the operation name (if available), and the
+        parameters used.
+
+        Examples
+        --------
+        >>> cf.print_operation_history()
+        1: normalize {}
+        2: low_pass_filter {'cutoff': 1000}
+        """
+        if not self.operation_history:
+            print("Operation history: <empty>")
+            return
+
+        print(f"Operation history ({len(self.operation_history)}):")
+        for i, record in enumerate(self.operation_history, start=1):
+            # record is expected to be a dict with at least a 'operation' key
+            op_name = record.get("operation") or record.get("name") or "<unknown>"
+            # Copy params for display - exclude the 'operation'/'name' keys
+            params = {k: v for k, v in record.items() if k not in ("operation", "name")}
+            print(f"{i}: {op_name} {params}")
+
+    def to_numpy(self) -> T:
+        """Convert the frame data to a NumPy array.
+
+        This method computes the Dask array and returns it as a concrete NumPy array.
+        The returned array has the same shape as the frame's data.
+
+        Returns
+        -------
+        T
+            NumPy array containing the frame data.
+
+        Examples
+        --------
+        >>> cf = ChannelFrame.read_wav("audio.wav")
+        >>> data = cf.to_numpy()
+        >>> print(f"Shape: {data.shape}")  # (n_channels, n_samples)
+        """
+        return self.data
+
+    def to_dataframe(self) -> "pd.DataFrame":
+        """Convert the frame data to a pandas DataFrame.
+
+        This method provides a common implementation for converting frame data
+        to pandas DataFrame. Subclasses can override this method for custom behavior.
+
+        Returns
+        -------
+        pd.DataFrame
+            DataFrame with appropriate index and columns.
+
+        Examples
+        --------
+        >>> cf = ChannelFrame.read_wav("audio.wav")
+        >>> df = cf.to_dataframe()
+        >>> print(df.head())
+        """
+        # Get data as numpy array
+        data = self.to_numpy()
+
+        # Get column names from subclass
+        columns = self._get_dataframe_columns()
+
+        # Get index from subclass
+        index = self._get_dataframe_index()
+
+        # Create DataFrame
+        if data.ndim == 1:
+            # Single channel case - reshape to 2D
+            df = pd.DataFrame(data.reshape(-1, 1), columns=columns, index=index)
+        else:
+            # Multi-channel case - transpose to (n_samples, n_channels)
+            df = pd.DataFrame(data.T, columns=columns, index=index)
+
+        return df
+
+    @abstractmethod
+    def _get_dataframe_columns(self) -> list[str]:
+        """Get column names for DataFrame.
+
+        This method should be implemented by subclasses to provide
+        appropriate column names for the DataFrame.
+
+        Returns
+        -------
+        list[str]
+            List of column names.
+        """
+        pass
+
+    @abstractmethod
+    def _get_dataframe_index(self) -> "pd.Index[Any]":
+        """Get index for DataFrame.
+
+        This method should be implemented by subclasses to provide
+        appropriate index for the DataFrame based on the frame type.
+
+        Returns
+        -------
+        pd.Index
+            Index for the DataFrame.
+        """
+        pass
+
     def _debug_info_impl(self) -> None:
         """Implement derived class-specific debug information"""
         pass
+
+    def _print_operation_history(self) -> None:
+        """Print the operation history information.
+
+        This is a helper method for info() implementations to display
+        the number of operations applied to the frame in a consistent format.
+        """
+        if self.operation_history:
+            print(f"  Operations Applied: {len(self.operation_history)}")
+        else:
+            print("  Operations Applied: None")
