@@ -365,7 +365,10 @@ class TestChannelProcessing:
 
             # _channel_metadata から ref を取得するケース
             frame = self.channel_frame
-            frame._channel_metadata = [mock.Mock(ref=0.5), mock.Mock(ref=1.0)]
+            frame._channel_metadata = [
+                ChannelMetadata(label="ch0", unit="", ref=0.5, extra={}),
+                ChannelMetadata(label="ch1", unit="", ref=1.0, extra={}),
+            ]
             result2 = frame.rms_trend()
             mock_create_op.assert_called_with(
                 "rms_trend",
@@ -391,6 +394,11 @@ class TestChannelProcessing:
         with mock.patch("wandas.processing.create_operation") as mock_create_op:
             mock_op = mock.MagicMock()
             mock_op.process.return_value = self.dask_data
+            # Mock get_metadata_updates to return updated sampling rate
+            hop_length = 256
+            mock_op.get_metadata_updates.return_value = {
+                "sampling_rate": self.sample_rate / hop_length
+            }
             mock_create_op.return_value = mock_op
 
             result = self.channel_frame.rms_trend(frame_length=1024, hop_length=256)
@@ -420,9 +428,15 @@ class TestChannelProcessing:
                         f"unit='Pa'のrefが一致しません: "
                         f"{res_meta.ref} != {base_meta.ref}"
                     )
-        assert getattr(result, "_channel_metadata", None) == getattr(
-            base, "_channel_metadata", None
-        )
+                # Check that labels are updated (not the same as base)
+                # Labels should now contain the operation name
+                if op_key:
+                    assert base_meta.label in res_meta.label, (
+                        f"Expected base label '{base_meta.label}' to be in "
+                        f"result label '{res_meta.label}'"
+                    )
+        # Note: We no longer check for exact equality of _channel_metadata
+        # because labels are now updated to reflect operations
         assert len(result.operation_history) == len(base.operation_history) + 1
         assert result.previous is base
 
@@ -560,8 +574,267 @@ class TestChannelProcessing:
         with mock.patch("wandas.processing.create_operation") as mock_create_op:
             mock_op = mock.MagicMock()
             mock_op.process.return_value = self.dask_data
+            # Mock get_metadata_updates to return target sampling rate
+            target_sr = 8000
+            mock_op.get_metadata_updates.return_value = {"sampling_rate": target_sr}
             mock_create_op.return_value = mock_op
-            result = self.channel_frame.resampling(target_sr=8000)
+            result = self.channel_frame.resampling(target_sr=target_sr)
             self._check_channel_frame_attrs(
                 result, self.channel_frame, hop_length=2, op_key="resampling"
             )
+
+
+class TestSamplingRateUpdates:
+    """Integration tests for sampling rate updates via metadata."""
+
+    def setup_method(self) -> None:
+        """Set up test fixtures."""
+        self.sample_rate = 44100
+        t = np.linspace(0, 1, self.sample_rate, endpoint=False)
+        signal = np.array([0.1 * np.sin(2 * np.pi * 440 * t)])
+        self.dask_data = _da_from_array(signal, chunks=-1)
+        self.frame = ChannelFrame(data=self.dask_data, sampling_rate=self.sample_rate)
+
+    def test_loudness_zwtv_updates_sampling_rate(self) -> None:
+        """Test that loudness_zwtv correctly updates sampling rate."""
+        loudness = self.frame.loudness_zwtv(field_type="free")
+
+        # Sampling rate should be updated to 500 Hz (2ms time steps)
+        assert loudness.sampling_rate == 500.0
+        assert loudness.sampling_rate != self.sample_rate
+
+    def test_rms_trend_updates_sampling_rate(self) -> None:
+        """Test that rms_trend correctly updates sampling rate."""
+        hop_length = 512
+        rms = self.frame.rms_trend(hop_length=hop_length)
+
+        # Sampling rate should be updated based on hop_length
+        expected_sr = self.sample_rate / hop_length
+        assert np.isclose(rms.sampling_rate, expected_sr)
+        assert rms.sampling_rate != self.sample_rate
+
+    def test_resampling_updates_sampling_rate(self) -> None:
+        """Test that resampling correctly updates sampling rate."""
+        target_sr = 16000
+        resampled = self.frame.resampling(target_sr=target_sr)
+
+        # Sampling rate should be updated to target_sr
+        assert resampled.sampling_rate == target_sr
+        assert resampled.sampling_rate != self.sample_rate
+
+    def test_operations_without_metadata_updates_preserve_sampling_rate(
+        self,
+    ) -> None:
+        """Test that operations without metadata updates preserve sampling rate."""
+        # Operations that don't change sampling rate
+        filtered = self.frame.low_pass_filter(cutoff=1000)
+        a_weighted = self.frame.a_weighting()
+        power_op = self.frame.power(exponent=2.0)
+
+        # Sampling rate should remain unchanged
+        assert filtered.sampling_rate == self.sample_rate
+        assert a_weighted.sampling_rate == self.sample_rate
+        assert power_op.sampling_rate == self.sample_rate
+
+    def test_chained_operations_with_sampling_rate_updates(self) -> None:
+        """Test chained operations that update sampling rate."""
+        # Chain operations: filter -> a_weighting -> rms_trend
+        hop_length = 512
+        result = (
+            self.frame.low_pass_filter(cutoff=5000)
+            .a_weighting()
+            .rms_trend(hop_length=hop_length)
+        )
+
+        # Final sampling rate should reflect rms_trend's update
+        expected_sr = self.sample_rate / hop_length
+        assert np.isclose(result.sampling_rate, expected_sr)
+
+
+class TestRoughnessOperations:
+    """Test roughness calculation operations."""
+
+    def setup_method(self) -> None:
+        """Set up test fixtures for each test."""
+        # Create a test signal (1 second at 44100 Hz)
+        self.sample_rate: float = 44100.0
+        duration: float = 1.0
+        t = np.linspace(0, duration, int(self.sample_rate * duration))
+
+        # Create a signal with modulated amplitude (roughness stimuli)
+        carrier_freq = 1000.0  # 1 kHz carrier
+        mod_freq = 70.0  # 70 Hz modulation (creates roughness)
+        signal = np.sin(2 * np.pi * carrier_freq * t) * (
+            1 + 0.5 * np.sin(2 * np.pi * mod_freq * t)
+        )
+
+        self.data: np.ndarray = signal.reshape(1, -1)  # 1 channel
+        self.dask_data: DaArray = _da_from_array(self.data, chunks=(1, 4410))
+        self.frame: ChannelFrame = ChannelFrame(
+            data=self.dask_data,
+            sampling_rate=self.sample_rate,
+            label="roughness_test",
+        )
+
+    def test_roughness_dw_basic(self) -> None:
+        """Test basic roughness_dw calculation."""
+        result = self.frame.roughness_dw(overlap=0.5)
+
+        # Check that result is a ChannelFrame
+        assert isinstance(result, ChannelFrame)
+
+        # Check that data shape is reduced (time-varying roughness)
+        # Note: mono signals are squeezed by ChannelFrame.data property to 1D
+        assert result.data.ndim in (1, 2)
+        n_time_points = (
+            result.data.shape[0] if result.data.ndim == 1 else result.data.shape[1]
+        )
+        original_samples = (
+            self.frame.data.shape[0]
+            if self.frame.data.ndim == 1
+            else self.frame.data.shape[1]
+        )
+        assert n_time_points < original_samples  # Reduced time points
+
+        # Check that sampling rate is updated
+        # For overlap=0.5 with 200ms windows, sampling rate should be ~10 Hz
+        assert result.sampling_rate < self.sample_rate
+        # Check operation history (accept both 'name' and 'operation' keys)
+        assert len(result.operation_history) == 1
+        first_op = result.operation_history[0]
+        op_name = first_op.get("name") or first_op.get("operation")
+        assert op_name == "roughness_dw"
+        assert first_op["params"]["overlap"] == 0.5
+
+    def test_roughness_dw_different_overlap(self) -> None:
+        """Test roughness_dw with different overlap values."""
+        result_overlap_0 = self.frame.roughness_dw(overlap=0.0)
+        result_overlap_05 = self.frame.roughness_dw(overlap=0.5)
+
+        # Higher overlap should result in more time points
+        n_time_0 = (
+            result_overlap_0.data.shape[0]
+            if result_overlap_0.data.ndim == 1
+            else result_overlap_0.data.shape[1]
+        )
+        n_time_05 = (
+            result_overlap_05.data.shape[0]
+            if result_overlap_05.data.ndim == 1
+            else result_overlap_05.data.shape[1]
+        )
+        assert n_time_05 > n_time_0
+
+        # Sampling rates should be different
+        assert result_overlap_05.sampling_rate > result_overlap_0.sampling_rate
+
+    def test_roughness_dw_validates_overlap(self) -> None:
+        """Test that roughness_dw validates overlap parameter."""
+        with pytest.raises(ValueError, match="overlap must be in"):
+            self.frame.roughness_dw(overlap=1.5)
+
+        with pytest.raises(ValueError, match="overlap must be in"):
+            self.frame.roughness_dw(overlap=-0.1)
+
+    def test_roughness_dw_spec_basic(self) -> None:
+        """Test basic roughness_dw_spec calculation."""
+        from mosqito.sq_metrics import roughness_dw as roughness_dw_mosqito
+
+        from wandas.frames.roughness import RoughnessFrame
+
+        result = self.frame.roughness_dw_spec(overlap=0.5)
+
+        # Check that result is a RoughnessFrame
+        assert isinstance(result, RoughnessFrame)
+
+        # Check dimensions
+        # Mono signal: (47, n_time), Multi-channel: (n_channels, 47, n_time)
+        assert result.data.ndim == 2  # (n_bark_bands, n_time) for mono
+        assert result.data.shape[0] == 47  # 47 Bark bands
+        assert result.data.shape[1] > 0  # Time points
+
+        # Check bark_axis
+        assert len(result.bark_axis) == 47
+        assert result.bark_axis[0] == pytest.approx(0.5, abs=0.1)
+        assert result.bark_axis[-1] == pytest.approx(23.5, abs=0.1)
+
+        # Check properties
+        assert result.n_bark_bands == 47
+        assert result.overlap == 0.5
+        assert len(result.time) == result.n_time_points
+        # Check operation history (accept both 'name' and 'operation' keys)
+        assert len(result.operation_history) == 1
+        first_op = result.operation_history[0]
+        op_name = first_op.get("name") or first_op.get("operation")
+        assert op_name == "roughness_dw_spec"
+
+        # Compare with MoSQITo direct calculation
+        computed_data = (
+            result.data.compute() if hasattr(result.data, "compute") else result.data
+        )
+        _, r_spec_direct, _, _ = roughness_dw_mosqito(
+            self.data[0], self.sample_rate, overlap=0.5
+        )
+        np.testing.assert_array_equal(
+            computed_data,
+            r_spec_direct,
+            err_msg="Specific roughness values differ from MoSQITo calculation",
+        )
+
+    def test_roughness_dw_spec_plot(self) -> None:
+        """Test that roughness_dw_spec plot method works."""
+        import matplotlib.pyplot as plt
+
+        result = self.frame.roughness_dw_spec(overlap=0.5)
+
+        # Should not raise an error
+        ax = result.plot()
+        assert ax is not None
+
+        plt.close("all")
+
+    def test_roughness_consistency(self) -> None:
+        """Test that roughness_dw and roughness_dw_spec are consistent."""
+        roughness = self.frame.roughness_dw(overlap=0.5)
+        roughness_spec = self.frame.roughness_dw_spec(overlap=0.5)
+
+        # Time points should match
+        n_time_roughness = (
+            roughness.data.shape[0]
+            if roughness.data.ndim == 1
+            else roughness.data.shape[1]
+        )
+        assert n_time_roughness == roughness_spec.n_time_points
+
+        # Total roughness should approximately equal 0.25 * sum(R_spec)
+        # R = 0.25 * sum(R_spec) according to Daniel & Weber
+        # roughness_spec.data: (47, n_time) for mono
+        total_from_spec = 0.25 * roughness_spec.data.sum(axis=0)
+        total_direct = roughness.data
+
+        # They should be close (allowing for numerical differences)
+        np.testing.assert_allclose(
+            total_direct.flatten(), total_from_spec.flatten(), rtol=0.1, atol=0.01
+        )
+
+    def test_roughness_multi_channel(self) -> None:
+        """Test roughness calculation with multi-channel signal."""
+        # Create 2-channel signal
+        data_2ch = np.vstack([self.data, self.data * 0.8])
+        dask_data_2ch: DaArray = _da_from_array(data_2ch, chunks=(1, 4410))
+        frame_2ch: ChannelFrame = ChannelFrame(
+            data=dask_data_2ch,
+            sampling_rate=self.sample_rate,
+            label="roughness_2ch",
+        )
+
+        # Test roughness_dw
+        roughness = frame_2ch.roughness_dw(overlap=0.5)
+        # Multi-channel data is NOT squeezed
+        assert roughness.data.ndim == 2
+        assert roughness.data.shape[0] == 2  # 2 channels
+
+        # Test roughness_dw_spec
+        roughness_spec = frame_2ch.roughness_dw_spec(overlap=0.5)
+        assert roughness_spec.data.ndim == 3
+        assert roughness_spec.data.shape[0] == 2  # 2 channels
+        assert roughness_spec.data.shape[1] == 47  # 47 Bark bands
