@@ -1,4 +1,5 @@
 # wandas/utils/__init__.py
 from .introspection import accepted_kwargs, filter_kwargs
+from .util import validate_sampling_rate
 
-__all__ = ["filter_kwargs", "accepted_kwargs"]
+__all__ = ["filter_kwargs", "accepted_kwargs", "validate_sampling_rate"]
